Move SimpleView to shared components. GenomeAssemblyAndAnnotationView --> SimpleView.
