--- src/proteomes/components/GenomeAssemblyAndAnnotationView.tsx
@@ -1,20 +1,0 @@
-import { FC } from 'react';
-import { ExternalLink } from 'franklin-sites';
-import { GenomeAssembly } from '../adapters/proteomesConverter';
-
-const GenomeAssemblyAndAnnotationView: FC<Partial<GenomeAssembly>> = ({
-  assemblyId,
-  source,
-  genomeAssemblyUrl,
-}) => {
-  if (!assemblyId) {
-    return null;
-  }
-  const text = `${assemblyId}${source ? ` from ${source}` : ''}`;
-  if (genomeAssemblyUrl) {
-    return <ExternalLink url={genomeAssemblyUrl}>{text}</ExternalLink>;
-  }
-  return <>{text}</>;
-};
-
-export default GenomeAssemblyAndAnnotationView;
