+++ src/shared/components/views/SimpleView.tsx
